+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-## Upcoming Release
+# Upcoming Release
 
 edbee.lib:
 
+- add #135, QAccessibleTextInterface support
+- add TextBuffer::emitTextChanged is extended with the oldText. SIGNAL void emitTextChanged( TextBufferChange* change, QString oldText = QString()); (WARNING: signal interface changes )
 - fix #133, Added Qt 6.3.0 compatibility
 - fix #132, Writing diacritics in Linux
 - ref #128, Build in supprort for virtual characters. Alternate position/cursor calculations via TextLayout